Check if testing user exist before adding
